// src/types/business.ts
export type Business = {
  id: number;
  type?: string;
  name?: string;
  street?: string;
  zip_code?: string;
  state?: string;
  qb_vendor_name?: string;
  qb_city_state_zip?: string;
  primary_phone?: string;
  backup_phone?: string;
  primary_email?: string;
  comments?: string;
  fax_phone?: string;
  contact_name?: string;
  website?: string;
  business_hours?: string;
  find_out?: string;
  onboarding_status?: string;
  join_date?: string;
  input_type_status?: string;
  vendor_type?: string;
  status?: string;
  pets_of_the_homeless_discount?: boolean;
  updated_by?: string;
  updated_date_time?: string;
  sync_to_qb?: boolean;
  veterinary?: boolean;
  ressoure?: boolean;
  food?: boolean;
  donation?: boolean;
  family_shelter?: boolean;
  wellness?: boolean;
  spay_neuter?: boolean;
  financial?: boolean;
  re_home?: boolean;
  er_boarding?: boolean;
  senior?: boolean;
  cancer?: boolean;
  dog?: boolean;
  cat?: boolean;
  fph_phone?: string;
  contact_phone?: string;
  web_notes?: string;
  internal_notes?: string;
  published?: boolean;
  shelter?: boolean;
  domestic_violence?: boolean;
  web_date_init?: string;
  ent_qb?: boolean;
  service_request?: boolean;
  inactive?: boolean;
  final_check?: boolean;
  created_by?: string;
  created_date?: string;
  city?: string;
  residential?: string;
  supply_request_status?: string;
};